Capture log output before docker stack is torn down
Previously this would attempt to grab the logs on a stack that had already been removed, so there would be nothing to read. This will now pipe the logs to a file and then read from that file in the event of any failures. Logs are written unconditionally so they can be checked manually in success cases if desired.