Add a PATH_PREFIX variable
This allows us to "easily" deploy Django to a path prefix but has a ton
of caveats and pitfalls that need to be addressed in subsequent commits.
Specifically: tests will not work with the current definition, nor will
our generated documentation. All of it assumes that the API is deployed
at the root of the URL. We can change it to use `BASE_URL`, but then
tests will fail if `PATH_PREFIX` is defined unless we switch to `reverse`
retrieve the paths in tests. Doing this, however, creates a ton of
circular imports.